pyproject.toml
# ---------------------------------------------------------------------------
# Pokédex Arcana — root workspace (uv)
#
# Cada paquete del monorepo (domain, shared, agents, infrastructure, api) es
# instalable y vive en su propio sub-pyproject. Aquí declaramos:
#   - Las dependencias externas comunes a todo el backend.
#   - Los miembros del workspace.
#   - Las herramientas (ruff, mypy, pytest).
#
# Comandos típicos:
#   uv sync                # instala TODO el workspace en .venv
#   uv run pytest          # corre la suite del dominio
#   uv run uvicorn api.main:app --reload --port 8000
# ---------------------------------------------------------------------------

[project]
name = "pokedex-arcana"
version = "0.1.0"
description = "Sistema multi-agente para responder preguntas Pokémon con datos verificables."
readme = "README.md"
requires-python = ">=3.11,<3.13"
license = { text = "MIT" }
authors = [{ name = "Pokédex Arcana Team" }]

dependencies = [
    # --- Paquetes locales del workspace ---------------------------------
    "pokedex-shared",
    "pokedex-domain",
    "pokedex-infrastructure",
    "pokedex-agents",
    "pokedex-api",

    # --- Web framework ---------------------------------------------------
    "fastapi>=0.115,<0.117",
    "uvicorn[standard]>=0.32,<0.36",
    "sse-starlette>=2.1,<3",
    "python-multipart>=0.0.20",

    # --- Modelado / validación ------------------------------------------
    "pydantic>=2.9,<3",
    "pydantic-settings>=2.5,<3",

    # --- LLMs ------------------------------------------------------------
    "groq>=0.11,<1",

    # --- Orquestación de agentes ----------------------------------------
    "langgraph>=0.2.50,<0.4",
    "langchain-core>=0.3.20,<0.4",
    "langchain-groq>=0.3.0,<0.4",
    "langchain-ollama>=0.3.0,<0.4",

    # --- Datos -----------------------------------------------------------
    "duckdb>=1.1,<2",
    "qdrant-client>=1.12,<2",
    "aiosqlite>=0.20,<0.22",
    "aiohttp>=3.10,<4",
    "httpx>=0.27,<1",
    "beautifulsoup4>=4.12,<5",
    "lxml>=5.3,<6",

    # --- Reportes --------------------------------------------------------
    "weasyprint>=63,<66",
    "markdown-it-py>=3.0,<4",

    # --- Observability / logs -------------------------------------------
    "structlog>=24.4,<26",
    "langfuse>=2.50,<3",

    # --- Utilidades ------------------------------------------------------
    "python-dotenv>=1.0,<2",
    "tenacity>=9.0,<10",
    "rich>=13.9,<14",
    "typer>=0.13,<1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3,<9",
    "pytest-asyncio>=0.24,<1",
    "pytest-cov>=5.0,<7",
    "ruff>=0.7,<1",
    "mypy>=1.13,<2",
    "respx>=0.21,<1",
]
eval = [
    "ragas>=0.2.10,<0.3",
    "datasets>=3.1,<4",
    "pandas>=2.2,<3",
]

# ---------------------------------------------------------------------------
# uv workspace
# ---------------------------------------------------------------------------

[tool.uv.workspace]
members = [
    "packages/shared",
    "packages/domain",
    "packages/infrastructure",
    "packages/agents",
    "apps/api",
]

[tool.uv.sources]
pokedex-shared = { workspace = true }
pokedex-domain = { workspace = true }
pokedex-infrastructure = { workspace = true }
pokedex-agents = { workspace = true }
pokedex-api = { workspace = true }

# ---------------------------------------------------------------------------
# Build (no se publica este root como wheel; sólo es agregador del workspace).
# ---------------------------------------------------------------------------

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
bypass-selection = true

# ---------------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------------

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra --strict-markers --strict-config --import-mode=importlib"
asyncio_mode = "auto"
testpaths = [
    "packages/domain/tests",
    "packages/agents/tests",
    "packages/infrastructure/tests",
    "apps/api/tests",
    "eval",
]
pythonpath = [
    "packages/shared/src",
    "packages/domain/src",
    "packages/infrastructure/src",
    "packages/agents/src",
    "apps/api/src",
]
markers = [
    "slow: pruebas que requieren red o servicios externos",
    "integration: pruebas contra Qdrant/Langfuse/PokéAPI reales",
]

# ---------------------------------------------------------------------------
# Ruff
# ---------------------------------------------------------------------------

[tool.ruff]
line-length = 100
target-version = "py311"
src = [
    "packages/shared/src",
    "packages/domain/src",
    "packages/infrastructure/src",
    "packages/agents/src",
    "apps/api/src",
]

[tool.ruff.lint]
select = [
    "E", "F", "W",      # pycodestyle / pyflakes
    "I",                 # isort
    "B",                 # bugbear
    "UP",                # pyupgrade
    "RUF",               # ruff-specific
    "SIM",               # simplify
    "PL",                # pylint subset
]
ignore = [
    "PLR0913",   # demasiados argumentos: domain calculator los necesita
    "PLR2004",   # magic numbers OK en fórmulas Gen IX (1.5, 2.0, etc.)
    "PLR0912",   # demasiadas ramas: DamageCalculator.calculate aplica 9 modificadores
    "PLR0915",   # demasiadas sentencias: idem
    "PLW0603",   # `global` aceptable para guards idempotentes (configure_logging)
    "RUF001",    # caracteres unicode ambiguos en strings (español + matemáticos: ×, →, —)
    "RUF002",    # idem en docstrings (intencional, lengua española)
    "RUF003",    # idem en comentarios
    "UP042",     # `class X(str, Enum)` es el patrón consagrado del codebase
    "SIM102",    # `if A: if B:` separado a veces es más legible
    "PLC0415",   # imports dentro de función justificados (evitar circular en main.py)
    "B008",      # `typer.Option(...)` y `Field(...)` son patrones idiomáticos en Typer/FastAPI/Pydantic
]

[tool.ruff.lint.per-file-ignores]
"**/tests/**" = ["PLR2004", "S101"]

# ---------------------------------------------------------------------------
# Mypy (estricto pero pragmático para hackathon)
# ---------------------------------------------------------------------------

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
disallow_untyped_decorators = false
plugins = ["pydantic.mypy"]
mypy_path = [
    "packages/shared/src",
    "packages/domain/src",
    "packages/infrastructure/src",
    "packages/agents/src",
    "apps/api/src",
]

[[tool.mypy.overrides]]
module = ["weasyprint.*", "qdrant_client.*", "langfuse.*", "ragas.*"]
ignore_missing_imports = true
